Changed how anchor rotated objects are drawn...
The position of the rectangle that has been rotated about an anchor
point is calculated in RectangleF (this has been so since
Dnawrkshp/mini2d@fa46af0311390064971a02c6ffd8713a1a5328df).
In Image::Draw(), if UseAnchor is set true then it will use the
calculated position from GetRotatedCenter() as the anchor and position
for Mini2D::DrawTexture().

This allows the image to be rotated [0-360] separately from the anchor
point rotation.