Rewrote TTF loading in Font class
Improved quality of drawn character bitmaps by loading the designed bitmap into the VRAM instead of cropping it by an arbitrary resolution and then scaling it to make it look normal.

Font::Load() functions use a pixel resolution specifying the pixel height to render the glyphs in. The larger the number, the more VRAM it takes up and the more accurate the character bitmap is.