Changed bpp of glyphs in the VRAM to 8 from 16
Previously, the A4R4G4B4 format was used. The grayscale glyph was divided by 4 and placed in the alpha channel. If the alpha was greater than 0, R4G4B4 were all set to 255. Not only is detail lost by truncating half of the bits, there are 12 unnecessary bits. Using the L8 texture format the grayscale is not downscaled, we use less bits per pixel, and there is less processing done per pixel (we can just copy the buffer directly into VRAM).